Add files via upload
Here I have to fix the pie charts, have to preprocess the data differently than min-max normalization, thinking maybe to change the negative to -1 and positive to 1, keeping all the data values in the rnage -1 to 1, inclusive. The logit function does not work here I think due to the 0-1 scale, it does not converge.